Rotate image with EXIF information
Using the [JavaScript-Load-Image](https://github.com/blueimp/JavaScript-Load-Image) library
the image orientation is first parsed from the uploaded file and the
used to create a CANVAS element with the correct orientation. This is
then encoded into a base64 data blob that can be used by `cropit` to
render the preview image as normal.